Rename repo to client-e2e-python-test-library
Drop the company name from the repo name. It's not a good practice.
