+++ .github/workflows/MCCE-Test-Linux.yml
@@ -72,6 +72,6 @@ jobs:
           cd $MCCE_ROOT_DIR
           pip install -r TESTS/pelion-e2e-python-test-library/requirements.txt
           pip install git+https://${{ secrets.ACCESS_TOKEN }}@github.com/PelionIoT/raas-pyclient.git#egg=raas-client
-          export PELION_CLOUD_API_KEY=${{ secrets.CLOUD_API_KEY }}
+          export CLOUD_API_KEY=${{ secrets.CLOUD_API_KEY }}
           pytest TESTS/pelion-e2e-python-test-library/tests/dev-client-tests.py --update_bin=__x86_x64_NativeLinux_mbedtls/Debug/mbedCloudClientExample.elf --local_binary=__x86_x64_NativeLinux_mbedtls/Debug/mbedCloudClientExample.elf --log-cli-level=DEBUG
       - run: echo "🍏 This job's status is ${{ job.status }}."
