+++ changelog.md
@@ -1,4 +1,10 @@
-# Pelion E2E Python Test Library Changelog
+# Client E2E Python Test Library Changelog
+
+## 0.3.0 2023-12-12
+- Rename the library to client-e2e-python-test-library.
+    - Rename `PELION_CLOUD_API_KEY` to `CLOUD_API_KEY`.
+    - Rename `PELION_CLOUD_API_GW` also to `CLOUD_API_GW`.
+- Update `requests` to latest (2.31.0).
 
 ## 0.2.10 2021-03-04
 - Update test doesn't depend on mbed bootloader anymore
