+++ client_test_lib/tools/client_runner.py
@@ -1,5 +1,7 @@
 """
 Copyright 2019-2020 Pelion.
+Copyright (c) 2023 Izuma Networks
+
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
 You may obtain a copy of the License at
@@ -16,7 +18,7 @@ import logging
 import queue
 import threading
 from time import time
-import pelion_test_lib.tools.utils as utils
+import client_test_lib.tools.utils as utils
 
 flog = logging.getLogger("ClientRunner")
 flog.setLevel(logging.DEBUG)
