+++ client_test_lib/tools/external_conn.py
@@ -1,6 +1,8 @@
 # pylint: disable=broad-except
 """
 Copyright 2019-2020 Pelion.
+Copyright (c) 2023 Izuma Networks
+
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
 You may obtain a copy of the License at
