+++ requirements.txt
@@ -3,5 +3,5 @@ mbed-ls==1.8.*
 pytest==6.2.*
 pytest-html
 pyserial
-requests==2.24.0
+requests==2.31.0
 ws4py
